presenter-view-interface.h: higher level interface
The presenter kanban-application.c is a higher level component
than the view kanban-window.c. Thus the dependencies should
run in the direction from kanban-window.c to kanban-application.c.
presenter-view-interface.h achieves this as an interface that
declares the view functions required and expecting them to exist,
rather than linking to those that do directly.

This change has a number of advantages:
 - Proper seperation of concerns between presenter and view
   + easier to test
   + More flexible, easier to change view and develop
 - A change to the view does not require presenter recompiling

Related todo:
 - explain architecture in docs
 - change folder names to make more concise and clear use purpose
 - Remove data that may change (e.g. version numbers) from higher
   to lower level components