kanban-cards: created external card model struct
Previously, in observer design pattern implementation, the Model
data was a stub struct in model-presenter-interface.h called
'KanbanModelData'. This has been replaced by a new struct,
'KanbanCard', externally in kanban-cards.h. The change_kanban()
function, previously static to model-presenter-interface.c,
is renamed, and called as a signal emitter, upon changes
to the KanbanCard model in kanban-cards.c.

This has the advantage of greater seperation between the
business logic and the presenter (interface). However, a
disadvantage is that the signal emitter function must be
declared in the interface header file even though it
should not be called by the presenter.
This is noted in a comment.

Relevant TODO:
 - Similarly, create proper ViewModel, and give it greater
   sepration from the interface implementation.

 - testing, asserts, test stub
 - Clean up linked list implementation (?)