kanban-list-store: added card-id hash table
Created GHashTable instance variable with card_id values as keys
and GSequenceIter pointers as values. The hash table is filled
during the _add_card() function and adds little overhead as
the GSequenceIter pointer is already held in a variable.

The advantage of this is that other functions, such as those to change
the content of existing cards, can access the GSequence iter in order
to change the underlying KanbanCardViewModel properties in O(1) time.
Without this, the GSequence would need to be traversed to find the
matching card-id item.

A call to g_list_model_items_changed() was also added to the
_add_card function to emit the GListModel "items-changed" signal.
This will eventually be neccessary upon the binding of the
GtkListBox GUI.

Relevant TODO:
 - Currently the keys (card_id values) are type gint. Thus the
   hash_func is g_direct_hash() and the key_equal_func is
   g_direct_equal(), using GINT_TO_POINTER() before calls.
   Eventually card_id will likely be changed to a string.
   This will require changes:
    * g_str_hash and g_str_equal
    * instead of calling GINT_TO_POINTER on keys, will need to call
      g_strdup to keep key in scope, and use g_hash_table_new_full()
      with g_free as the key_destroy_func.