kanban-list-store: changed params to struct
_change_column() _change_content() and _move_card() funcs now
use KanbanCard * struct as parameter, instead of the individual
member variables that are relevant to them. This is cleaner, clearer,
and should make it require less changes should this struct change
in the future.

This mirrors the change already made to the _new_card() function.

Change requires changes to declarations in header file, and to function
calls in model-observer.c.

Next step is to implement the functions acting on cards (columns TODO).