kanban-list-box: new barebones type for column GUI
kanban-list-box files create a GObject derived from GtkListBox.
This is intened to be bound to kanban-list-store in order to act
as the GUI for the column / card list stored there, via the
GListModel interface.

Currently, the files just contain the bare bones boilerplate
required by GTK.

Source file added to Makefile, but not yet called by any other
files.