kanban-list-box: bind model to display cards
gtk_list_box_bind_model() used to bind viewmodel of cards in
column in kanban-list-store to kanban-list-box. Some points to
note:
 - Cast the KanbanListBox object to GtkListBox rather than using
   GtkListBox parent_instance so that the derived object can be used
   as GtkListBox can in terms of displaying it as a widget.
 - Had to do the binding in _constructed() method rather than
   _init() because object properties are initialized after _init()
   call and so required "card-data" was NULL.

kanban-window altered to make sure widget is visible - this crude
method will need a proper implementation using .ui files.

Changes to other files trival or printing debugging messages.
These will need to be removed. An exception to this is a bug
found in kanban-card-viewmodel _update_content() function
where the content was being added to the heading.

Relevant TODO:
 - Remove debugging messages added and clean up code
 - Properly implement positioning of KanbanListBox in KanbanWindow
 - Proper implementation of card design using .ui files
 - Connect cards created to "notify" signals of corresponding
   KanbanCardViewModel to update heading and content upon changes
   (that don't emit "items-changed" signal of GListModel if not moved).

 - Implement multiple columns
 - Implement _change_column() and _move_card() functions in
   kanban-list-store

 - Implement Drag-and-Drop
 - Testing