meson.build: created meson build system
Single meson.build file created at root dir of project.
compile with:
$ meson builddir
$ cd builddir
$ ninja
The first time. After that simply:
$ cd builddir && ninja
The executable is in the builddir directory.

builddir/ added to .gitignore.

Note that the name arguement (1st parameter) to gnome.compile_resources
appears to just be used to prefix the variables defined in the
source files created (similar to c_name). This may become important
if multiple .gresources.xml files are created, but is not of
concern currently.

Related TODO:
 - Investigate -fpie warning of meson builddir
 - Perhaps add comments to justify flags used etc.
 - Remove old Makefile once depreciated