meson.build: consise test definition and cleanup
Each test is now added to meson using a single function
call, allowing them to be added more quickly and consisely.

As the sources required for each test are listed individually,
the seperatation of the main.c file added to the root meson.build
file upon testing was removed, for cleanness.