kanban-list-store: added column-id property
First step towards implementing multiple columns. List of
column names in KanbanListStore removed and replaced with
single column-id, as each KanbanListStore object is to represent
a single column of a kanban board.

Added test to check retrieval of property in
kanban-list-store-tests.cpp.

Changed definition of initialize_viewmodel() function to create object
in kanban-list-store, to have column id as a parameter. This function
will need changing names when mutliple columns implemented. This change
required changing the call in kanban-application.c and the declaration
in kanban-list-store.h and the stub definition in
application-entry-tests.cpp.

All tests pass inc. valgrind.