kanban-data: renamed KanbanCard to KanbanData
KanbanData struct used to exchange data from observers
to presenters via the observer pattern. This struct is
used to communicate info on columns as well as cards, so
the name KanbanCard was not appropriate.

Required changes to lots of files. However, struct in model
should (eventually) be very stable, so not a major architectural
issue.