kanban-application: moved files to root src/ dir
kanban-application is essentially the 'main' of the application.
It deals with calling the initialization of the presenters and
views, as well as IO with the OS and user input (ie. -V option).
Thus kanban-application moved to src/ directory to show that it
is not a presenter, but a low level detail. This is important,
as kanban-application can now be used for simple dependency
injection of the view into the presenters /viewmodel object(s).

presenter-view-interface.h consequently deleted, and the declaration
of initialize_view() added to kanban-window, where it is defined.
kanban-application is allowed to depend on kanban-window (or anything
else!).

Changes propaged to dependent tests, and meson.build files.

Summary: the purpose of this commit is to (start to) use kanban-application
pragmatically, to initialize the various components of the application,
as the 'main' component.

All tests pass, inc valgrind.