kanban-column-store: add view observer property
Property of type KanbanColumnViewer added to KanbanColumnStore.
This is for dependency injection. The view implementing the
KanbanColumnViewerInterface (to be kanban-grid) will be passed
by kanban-application as a arguement to the call to
kanban_column_store_new().

Currently, kanban-application simply passes NULL as arguement.

Use/faking of kanban_column_store_new updated to
application-entry-tests and kanban-column-store-tests, and
kanban-application.

All tests pass inc. valgrind.

Next steps are to inject kanban-grid in kanban-application,
and implement the event firing in kanban-column-store. Once
this is done, much cleanup of #include's can be done.