kanban-list-box: use GtkTextView for content
kanban-list-box now uses the new GtkTextBuffer implementation
of the content, defined in kanban-card-viewmodel. Properties
set to not be (directly) editable, and to not show cursor
(though this does not seem to work). This implementation,
successfully binds the content to the presenters, and is
updated when the content changes.

All tests pass inc. valgrind.

Points to consider:
 - Implement header the same way?
 - Cannot apply styles to GtkTextView, only via tags to GtkTextBuffer.
   GtkLabel etc. can (as copy text).