kanban-grid: add custom CSS
"board.css" file created to hold custom CSS rules. This file
was loaded as a GResource via "kanbna-app.gresrouce.xml" file and
activated in kanban-grid.

The CSS is intended to be done via style classes. These were added
to the "card.ui" file e.g.:

<object class="GtkClassName">
  <style>
    <class name="css-class-name"/>
  </style>
</object>

Note that even in the '*' selector is used in the CSS, changes are
limited to kanban-app only.

This allows a modular styling of view elements independently of
implementation. This solves the problem of style 'tags' being
applied to GtkTextBuffer, a presenter. As such the
kanban-card-viewmodel heading can now be implemented the same way
as the content. (The heading is significant as it is highly likely
the application will want to add default styling to it - e.g.
make bold.)

Relevant TODO:
 - implement header as GtkTextBuffer etc.
 - make UI styling attractive

Future TODO:
 - Add heading property as GtkTextBuffer to presenters/kanban-list-store
   and create interface with GListModel prerequisite to use instead of
   GListModel.
 - Create column.ui for use with kanban-grid, displaying column
   heading.
 - Create board.ui for use with kanban-grid, to hold the different
   columns