column.ui: add UI definition for column widgets
column.ui is to be used as a template for column-list-box,
which will derive from GtkWidget instead of GtkListBox. This
change will bring the following benefits:

 - Provides mechanism to display a column name
 - Provides clean, comparmentatised way to seperate column GUI
   structure/appearance/properties from the required logic.

New .ui file added to existing .gresource.xml file so it can
be used as a resource. Project successfully builds.

TODO: as has been previously recognised, file/class names will
need to be changed to make architecture/function clearer.