kanban-window: _add_board() for kanban-application
Use of gtk_container_add() in activate vfunc of
kanban-application was not clean, as initializations
were mixed with the layout of the GUI window. Thus
abstracted with new kanban_window_add_board() function.

Erroneous use of gtk_window_present() function also removed.

New function stubbed for use in application-entry-tests.

All tests pass and program runs as expected.

This is a continuation of cleanup of kanban-application.