kanban-column-viewmodel: _move_card tests & bugfix
Various tests for kanban_column_viewmodel_move_card() added
to kanban-column-viewmodel-tests.

BUGFIX where new_position needs to incremented if moving within
same column to lower position (as card being moved should not
be counted when finding the position to move to in this case).

Note the only remaining public function of kanban-column-viewmodel
without unit tests, the _get_sequence() function, will not be
given unit tests. This is because it is a simple getter used
only by kanban-board-presenter, and to unit test it would be
to test implementation rather than behaviour.

All tests pass inc. valgrind and clang. Program runs as expected.