column-view: fix memory leak & add card test
In function to create card widget in binding between GtkListBox
and GListModel, needed to unref the GtkBuilder object and ref
the Widget retrieved. This achives the transfer-full stated in
the docs.

Test added to check card widget added to view when card added in
viewmodel. Actual viewmodel code added to meson to achieve this.

All tests pass, inc. valgrind, clang and release.