kanban-column-view: add _get_heading() test func
Added function to head text of heading stored in a card
widget at a given priority (or NULL if not found). The char*
returned needs to be freed. This function is only for use in
testing, and is not visible to the compiler unless
TESTING_ONLY_ACCESS defined.

Relevant TODO:
 - A function was also created to find widgets by GtkBuildable name.
   This would better be moved to a new 'utils' file.
 - Add test of _edit_card to presenters-views-integration-tests
   using new function
 - There are other common tasks in tests, such as counting the
   number of cards, that could also better be replaced with
   testing only functions in kanban-column-view.