kanban-utils: new file; add _find_widget_by_name()
kanban-utils file created to hold general purpose functions that
do not belong to any particlar GObject or architectural unit,
for the purposes of cleaner, DRYer code.

kanban_utils_find_widget_by_name() function moved from
kanban-column-view. It will recursively find the child of a given widget
with a particular name (given as an 'id' attribute in GtkBuilder
UI definitions).

Given that the functions in kanban-utils are simply helper functions,
moved from other files, and not behaviours of the program itself,
it is not intended to be unit tested. Conversely, it would be
appropriate to include the source file in any unit test
neccessary.

All tests pass, inc. valgrind, clang and release.