kanban-board-view: add column_id->widget HashTable
GHashTable added to KanbanBoardView definition with the
"column-id" property value as the key and the column
widget created in _add_column() as the value. This
hash table is filled during calls to _add_column().

The purpose of this hash table is to be used in
implementing TASK_MOVE_COLUMN functionality.

All tests pass, inc. valgrind, clang and release.