meson: fetch & build gtest if not present
Meson will now build the gtest dependency from source,
fetching it as a subproject (added to gitignore) using
meson wrap. This will allow the tests to run in ubuntu
under Travis-CI, where gtest is not in the repos as a linkable
library.

Note that compiling from source using the same flags as the
project is google's reccomended way of using gtest.

Made valgrind flags verbose, altered flag not recognised in
valgrind 3.13 in ubuntu repos so will cause error.